Refactor to reuse code for opening popup window
