+++ src/Router.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import {BrowserRouter as Router, Route, Switch,} from 'react-router-dom';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Login from "./components/auth/Login";
 import Register from "./components/auth/Register";
-import {PrivateRoute, RedirectMain} from "./components/auth/Auth";
+import { PrivateRoute, RedirectMain } from "./components/auth/Auth";
 
 import Mine from "./views/IntegrationsMine";
 import Account from "./views/Account";
