+++ src/components/Instance.js
@@ -4,14 +4,16 @@ import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
 import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import Button from '@material-ui/core/Button';
-import {withTheme} from "@material-ui/core/styles/index";
+import { withTheme } from "@material-ui/core/styles/index";
 import React from 'react';
 import Loading from './Loading';
-import {get} from 'lodash';
+import { get } from 'lodash';
+
+import { openConfigWindow } from '../lib/configWindow';
 
 import {
-    updateSolutionInstanceConfig,
     updateSolutionInstance,
+    updateSolutionInstanceConfig,
 } from '../api/solutions';
 
 export class Instance extends React.PureComponent {
@@ -22,57 +24,7 @@ export class Instance extends React.PureComponent {
     };
 
     onClickConfigure = () => {
-        // Must open window from user interaction code otherwise it is likely
-        // to be blocked by a popup blocker:
-        const configWindow = window.open(
-            undefined,
-            '_blank',
-            'width=500,height=500,scrollbars=no'
-        );
-
-        // Listen to popup messages
-        let configFinished = false;
-        let configCanceled = false;
-        const onmessage = e => {
-            if (e.data.type === 'tray.configPopup.error') {
-                // Handle popup error message
-                alert(`Error: ${e.data.err}`);
-            }
-            if (e.data.type === 'tray.configPopup.cancel') {
-                // Handle popup canceled message
-                configCanceled = true;
-                configWindow.close();
-            }
-            if (e.data.type === 'tray.configPopup.finish') {
-                // Handle popup finish message
-                configFinished = true;
-                configWindow.close();
-            }
-        }
-        window.addEventListener('message', onmessage);
-
-        // Check if popup window has been closed before finishing the configuration.
-        // We use a polling function due to the fact that some browsers may not
-        // display prompts created in the beforeunload event handler.
-        const CHECK_TIMEOUT = 1000;
-        const checkWindow = () => {
-            if (configWindow.closed) {
-                // Handle popup closing
-                if (configFinished) {
-                    console.log('Configuration finished');
-                }
-                else if (configCanceled) {
-                    console.log('Configuration canceled');
-                }
-                else {
-                    alert('Configuration not finished');
-                }
-                window.removeEventListener('message', onmessage);
-            } else {
-                setTimeout(checkWindow, CHECK_TIMEOUT);
-            }
-        };
-        checkWindow();
+        const configWindow = openConfigWindow();
 
         updateSolutionInstanceConfig(this.props.id).then(({body}) => {
             // After we generate the popup URL, set it to the previously opened
