+++ src/components/Nav.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {withStyles} from '@material-ui/core/styles';
+import { withStyles } from '@material-ui/core/styles';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
@@ -11,7 +11,7 @@ import PlugIcon from '@material-ui/icons/SettingsInputComponent';
 import ExpandLess from '@material-ui/icons/ExpandLess';
 import ExpandMore from '@material-ui/icons/ExpandMore';
 import CircleIcon from '@material-ui/icons/FiberManualRecord';
-import {Link} from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 const styles = theme => ({
     root: {
