+++ src/views/SolutionsDiscover.js
@@ -11,6 +11,7 @@ import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import Loading from '../components/Loading';
 
+import { openConfigWindow } from '../lib/configWindow';
 import { listSolutions, createSolutionInstance } from '../api/solutions';
 
 export class DiscoverIntegrations extends React.PureComponent {
@@ -39,57 +40,7 @@ export class DiscoverIntegrations extends React.PureComponent {
     }
 
     onUseWorkflowClick(id, name) {
-        // Must open window from user interaction code otherwise it is likely
-        // to be blocked by a popup blocker:
-        const configWindow = window.open(
-            undefined,
-            '_blank',
-            'width=500,height=500,scrollbars=no',
-        );
-
-        // Listen to popup messages
-        let configFinished = false;
-        let configCanceled = false; 
-        const onmessage = e => {
-            if (e.data.type === 'tray.configPopup.error') {
-                // Handle popup error message
-                alert(`Error: ${e.data.err}`);
-            }
-            if (e.data.type === 'tray.configPopup.cancel') {
-                // Handle popup canceled message
-                configCanceled = true;
-                configWindow.close();
-            }
-            if (e.data.type === 'tray.configPopup.finish') {
-                // Handle popup finish message
-                configFinished = true;
-                configWindow.close();
-            }
-        }
-        window.addEventListener('message', onmessage);
-
-        // Check if popup window has been closed before finishing the configuration.
-        // We use a polling function due to the fact that some browsers may not
-        // display prompts created in the beforeunload event handler.
-        const CHECK_TIMEOUT = 1000;
-        const checkWindow = () => {
-            if (configWindow.closed) {
-                // Handle popup closing
-                if (configFinished) {
-                    console.log('Configuration finished');
-                }
-                else if (configCanceled) {
-                    console.log('Configuration canceled');
-                }
-                else {
-                    alert('Configuration not finished');
-                }
-                window.removeEventListener('message', onmessage);
-            } else {
-                setTimeout(checkWindow, CHECK_TIMEOUT);
-            }
-        }
-        checkWindow();
+        const configWindow = openConfigWindow();
 
         createSolutionInstance(id, name).then(({body}) => {
             // After we generate the popup URL, set it to the previously opened
